change type decimal to bigInteger

+++ database/migrations/2018_12_28_074455_create_transactions_table.php
@@ -18,7 +18,7 @@ class CreateTransactionsTable extends Migration
             $table->integer('user_id')->unsigned();
             $table->integer('type_id')->unsigned();
             $table->integer('category_id')->unsigned();
-            $table->decimal('amount');
+            $table->bigInteger('amount');
             $table->string('note');
             $table->date('date');
             $table->string('user');
